+++ Firebase/functions/src/fcm/schedule.ts
@@ -2,6 +2,7 @@ import { onSchedule } from "firebase-functions/v2/scheduler";
 import { getFunctions } from "firebase-admin/functions";
 import * as admin from "firebase-admin";
 import * as logger from "firebase-functions/logger";
+import { addDays, getZonedParts, zonedDateTimeToUTC } from "../common/date";
 import { normalizeError } from "../common/error";
 import { FirestorePath } from "../common/firestorePath";
 import { resolveTimeZone } from "./shared";
@@ -10,14 +11,6 @@ const LOCATION = "asia-northeast3";
 const DEFAULT_HOUR = 9;
 const DEFAULT_MINUTE = 0;
 const MINUTE_INTERVAL = 5;
-
-type ZonedDateParts = {
-    year: number;
-    month: number;
-    day: number;
-    hour: number;
-    minute: number;
-};
 
 // 사용자별 설정 시간에 맞춘 내일 마감 Todo 알림 작업 큐 적재
 export const scheduleTodoReminder = onSchedule({
@@ -143,88 +136,3 @@ export const scheduleTodoReminder = onSchedule({
         }
     }
 );
-
-// 지정한 타임존 기준 연월일시분 값 추출
-function getZonedParts(date: Date, timeZone: string): ZonedDateParts {
-    const parts = new Intl.DateTimeFormat("en-US", {
-        timeZone,
-        year: "numeric",
-        month: "2-digit",
-        day: "2-digit",
-        hour: "2-digit",
-        minute: "2-digit",
-        hour12: false
-    }).formatToParts(date);
-
-    const byType = (type: string): number => {
-        const found = parts.find((part) => part.type === type)?.value;
-        return Number(found);
-    };
-
-    return {
-        year: byType("year"),
-        month: byType("month"),
-        day: byType("day"),
-        hour: byType("hour"),
-        minute: byType("minute")
-    };
-}
-
-// GMT 오프셋 문자열의 분 단위 오프셋 값 변환
-function parseShortOffsetToMinutes(shortOffset: string): number {
-    if (shortOffset === "GMT" || shortOffset === "UTC") return 0;
-    const match = shortOffset.match(/^GMT([+-])(\d{1,2})(?::(\d{2}))?$/);
-    if (!match) return 0;
-
-    const sign = match[1] === "-" ? -1 : 1;
-    const hour = Number(match[2]);
-    const minute = Number(match[3] ?? "0");
-    return sign * (hour * 60 + minute);
-}
-
-// 특정 UTC 시점의 타임존 오프셋 분 단위 계산
-function getOffsetMinutesAt(utcDate: Date, timeZone: string): number {
-    const parts = new Intl.DateTimeFormat("en-US", {
-        timeZone,
-        timeZoneName: "shortOffset"
-    }).formatToParts(utcDate);
-
-    const offset = parts.find((part) => part.type === "timeZoneName")?.value ?? "GMT";
-    return parseShortOffsetToMinutes(offset);
-}
-
-// 타임존 기준 로컬 날짜 시간의 UTC Date 변환
-function zonedDateTimeToUTC(
-    year: number,
-    month: number,
-    day: number,
-    hour: number,
-    minute: number,
-    timeZone: string
-): Date {
-    const localAsUTC = Date.UTC(year, month - 1, day, hour, minute, 0, 0);
-    let utcMs = localAsUTC;
-
-    // DST 경계 시 오프셋이 바뀔 수 있어 2회 보정
-    for (let i = 0; i < 2; i += 1) {
-        const offsetMinutes = getOffsetMinutesAt(new Date(utcMs), timeZone);
-        utcMs = localAsUTC - offsetMinutes * 60 * 1000;
-    }
-
-    return new Date(utcMs);
-}
-
-// 주어진 로컬 날짜에 대한 일 수 가산 결과 반환
-function addDays(year: number, month: number, day: number, value: number): {
-    year: number;
-    month: number;
-    day: number;
-} {
-    const utcDate = new Date(Date.UTC(year, month - 1, day, 0, 0, 0, 0));
-    utcDate.setUTCDate(utcDate.getUTCDate() + value);
-    return {
-        year: utcDate.getUTCFullYear(),
-        month: utcDate.getUTCMonth() + 1,
-        day: utcDate.getUTCDate()
-    };
-}
